requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-filter>=2.4,<3.0
drf-spectacular>=0.17,<1.0
Pillow>=8.0,<9.0
psycopg2-binary>=2.9,<3.0
python-magic>=0.4,<1.0
python-dotenv>=0.19,<1.0
gunicorn>=20.1,<21.0